# Conflicts:
#	data/CONFIG/force_control/kinematic_solver_config.json
#	packages/cobotsys/CMakeLists.txt
#	packages/plugin_library/KinematicSolver/CMakeLists.txt
#	packages/plugin_library/KinematicSolver/orocos_kdl/CMakeLists.txt
#	packages/plugin_library/KinematicSolver/orocos_kdl/models/kukaLWR_DHnew.cpp
#	packages/plugin_library/KinematicSolver/orocos_kdl/src/CMakeLists.txt
#	packages/plugin_library/KinematicSolver/src/KinematicSolver.cpp
#	packages/plugin_library/KinematicSolver/src/KinematicSolver.h
#	packages/plugin_library/SimpleUi/CMakeLists.txt
#	packages/plugin_library/URRealTimeDriver/CMakeLists.txt

+++ packages/cobotsys/src/cobotsys_message.cpp
@@ -18,7 +18,7 @@ MessageEncoder::~MessageEncoder() {
 }
 
 Message MessageEncoder::genStringMessage(const std::string& msg) {
-    return genMessage(QByteArray(msg.c_str(), msg.size() + 1), MessageType::String);
+    return genMessage(QByteArray(msg.c_str(), (int)(msg.size() + 1)), MessageType::String);
 }
